Adapt to latest z3-ml.

+++ src/llvm2smt.ml
@@ -125,7 +125,7 @@ module Init (ZZ3 : ZZ3_sigs.S) (SMTg : module type of Smt_graph.Make(ZZ3))= stru
           T.rat (Q.of_float @@ Option.get value)
       | Int, ConstantInt ->
           let value = int64_of_const llv in
-          T.int (Z.of_int64 @@ Option.get value)
+          T.bigint (Z.of_int64 @@ Option.get value)
       | Bool, ConstantInt ->
           let value = int64_of_const llv in
           T.bool (0L <> Option.get value)
